github_app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
anthropic>=0.78.0
python-dotenv>=1.0.0
PyJWT[crypto]>=2.9.0
cryptography>=44.0.0
httpx>=0.28.0
